requirements.txt
langchain
openai
pandas
streamlit
python-dotenv
